fixed the reviews on mobile screen
I am starting small first so I only adjusted for the mobile screen, no
screenlisteners yet.
What I did: adding to the review divs so we have a nice flexbox so the
alignment of the name, date, rate, and comment is as in the mock up.
Size difference and background still have to be done.